Bug 1921925 - Factor out weather suggestions into their own file and move data from record to attachment
