+++ components/suggest/src/config.rs
@@ -5,7 +5,7 @@
 
 use serde::{Deserialize, Serialize};
 
-use crate::rs::{DownloadedGlobalConfig, DownloadedWeatherData};
+use crate::rs::DownloadedGlobalConfig;
 
 /// Global Suggest configuration data.
 #[derive(Clone, Default, Debug, Deserialize, Serialize, PartialEq, Eq, uniffi::Record)]
@@ -25,12 +25,4 @@ impl From<&DownloadedGlobalConfig> for SuggestGlobalConfig {
 #[derive(Clone, Debug, Deserialize, Serialize, PartialEq, Eq, uniffi::Enum)]
 pub enum SuggestProviderConfig {
     Weather { min_keyword_length: i32 },
-}
-
-impl From<&DownloadedWeatherData> for SuggestProviderConfig {
-    fn from(data: &DownloadedWeatherData) -> Self {
-        Self::Weather {
-            min_keyword_length: data.weather.min_keyword_length,
-        }
-    }
 }
